+++ CHANGELOG.md
@@ -20,6 +20,8 @@ All notable changes to this project will be documented in this file.
 
 ### 2.2.4
 - Add colors to embed responses
+- Fixed bugs with `m-animesearch` and `m-resetapi`
+- Minor style tweaks
 
 ## 2.1 - Michelle API (2019-03-22)
 - Introduced improved Michelle API. API is modular and modules can be added to the API folder.